pyproject.toml
[tool.poetry]
name = "learn-python"
version = "0.1.0"
description = "Vidhya's practice repository for learning python"
authors = ["Vidhya Chidambaram <vidhya.chidambaram@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pytest = "^9.0.2"
matplotlib = "^3.10.8"
seaborn = "^0.13.2"
ipywidgets = "^8.1.8"
pandas-stubs = "^2.3.3.251219"
poetry-core = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
